1.5.2

+++ ithemelandco-woo-report/class/setting_report.php
@@ -1165,6 +1165,12 @@ if (isset($_POST["update_settings"])) {
 
     //	print_r($_POST);
 
+    // Check nonce before processing any data
+    if (!isset($_POST['it_report_nonce_field']) || !wp_verify_nonce($_POST['it_report_nonce_field'], 'it_report_nonce_action')) {
+        // Exit if nonce is invalid or missing
+        wp_die(__('Nonce verification failed', 'ithemelandco-woo-report'));
+    }
+
     // Do the saving
     if (!in_array(__IT_REPORT_WCREPORT_FIELDS_PERFIX__ . 'set_default_fields', $_POST)) {
         delete_option(__IT_REPORT_WCREPORT_FIELDS_PERFIX__ . 'set_default_fields');
